+++ src/jk_sysinfo/entity/__init__.py
@@ -1,12 +1,10 @@
+﻿
 
 
-
-__version__ = "0.2021.12.30"
+__version__ = "0.2022.1.3"
 
 
 
 from .DriveInfo import DriveInfo
 from .DrivePartitionInfo import DrivePartitionInfo
-
-